requirements.txt
Django
gunicorn
django-heroku
django-allauth
Pillow
django-filter